+++ ui/src/common/index.js
@@ -25,7 +25,7 @@ export * from './LoveButton'
 export * from './Title'
 export * from './SongBulkActions'
 export * from './useAlbumsPerPage'
-export * from './useGetHandleArtistClick.jsx'
+export * from './useGetHandleArtistClick'
 export * from './useInterval'
 export * from './useResourceRefresh'
 export * from './useToggleLove'
@@ -39,3 +39,4 @@ export * from './useSelectedFields'
 export * from './ToggleFieldsMenu'
 export * from './QualityInfo'
 export * from './formatRange.js'
+export * from './playlistUtils.js'
